chore: move files, add tooltip
