+++ packages/paste-token-designer/components/preview/AnchorPreview.js
