+++ packages/paste-token-designer/components/preview/AvatarPreview.js
