+++ packages/paste-token-designer/components/preview/MenuPreview.js
