+++ packages/paste-token-designer/components/preview/PopoverPreview.js
@@ -4,9 +4,9 @@ import {Text} from '@twilio-paste/core/text';
 const PopoverPreview = () => {
   return (
     <PopoverContainer baseId="popover-example" placement="right-start" visible>
-      <PopoverButton variant="primary">✊ Action</PopoverButton>
+      <PopoverButton variant="primary">Open popover</PopoverButton>
       <Popover aria-label="Popover">
-        <Text as="span">Black Lives Matter. We stand with the Black community</Text>
+        <Text as="span">Black Lives Matter. We stand with the Black community.</Text>
       </Popover>
     </PopoverContainer>
   );
