+++ packages/paste-token-designer/components/preview/TablePreview.js
