+++ packages/paste-token-designer/components/preview/TooltipPreview.js
@@ -1,0 +1,12 @@
+import {Button} from '@twilio-paste/core/button';
+import {Tooltip} from '@twilio-paste/core/tooltip';
+
+const TooltipPreview = () => {
+  return (
+    <Tooltip text="Welcome to Paste!" placement="right" visible>
+      <Button variant="primary">Open tooltip</Button>
+    </Tooltip>
+  );
+};
+
+export {TooltipPreview};
